requirements.txt
Telethon==1.34.0
pyyaml>=6.0.2